backend/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
SQLAlchemy==2.0.21
PyJWT==2.8.0
bcrypt==4.0.1
python-dotenv==1.0.0
Flask-CORS==4.0.0
deepface==0.0.75
groq==0.4.1
opencv-python==4.8.0.76
requests==2.31.0
celery==5.3.1
redis==5.0.0
Pillow==10.0.0
numpy==1.24.3
scipy==1.11.2
protobuf==4.24.1
